--- FinanceDataMigrationApi/V1/UseCase/Interfaces/IAssetSaveToSqlUseCase.cs
@@ -1,10 +1,0 @@
-using System.Threading.Tasks;
-using System.Xml.Linq;
-
-namespace FinanceDataMigrationApi.V1.UseCase.Interfaces
-{
-    public interface IAssetSaveToSqlUseCase
-    {
-        public Task<int> ExecuteAsync(string lastHint, XElement xml);
-    }
-}
